:recycle: Seperate sub functions as internal function
