+++ utils/.internal/filterFuncProviders.js
@@ -1,0 +1,8 @@
+const filterFunc = (provider, filter, exclude) => {
+  if (!filter) return true
+
+  const filterIncludes = filter.includes(provider.provider_name)
+  return exclude === true ? !filterIncludes : filterIncludes
+}
+
+module.exports = filterFunc
