+++ utils/filterProviderKeys.js
@@ -1,12 +1,7 @@
+const filterFunc = require('./.internal/filterFuncProviderKeys')
+
 const filterProviderKeys = (keys, filter) => {
   if (!filter) return keys
-
-  const filterFunc = (key, filter, exclude) => {
-    if (!filter) return true
-
-    const filterIncludes = filter.includes(key)
-    return exclude ? !filterIncludes : filterIncludes
-  }
 
   return keys
     .filter(key => filterFunc(key, filter.include))
